rest-framework/requirements.txt
django==5.0.6
djangorestframework
pyyaml
requests
django-cors-headers
